pyproject.toml
[project]
name = "tagiato"
dynamic = ["version"]
description = "CLI nástroj pro automatické přidání GPS souřadnic a AI-generovaných popisků k JPEG fotografiím"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Pavel Mica" }
]
keywords = ["photos", "gps", "exif", "ai", "claude", "geolocation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "typer>=0.9.0",
    "piexif>=1.1.3",
    "Pillow>=9.0.0",
    "requests>=2.28.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
tagiato = "tagiato.cli.main:app"

[project.urls]
Homepage = "https://github.com/pavelmica/tagiato"
Repository = "https://github.com/pavelmica/tagiato"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "VERSION"
pattern = "(?P<version>.+)"

[tool.hatch.build.targets.wheel]
packages = ["src/tagiato"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/tagiato"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
